+++ .gitignore
@@ -1,1 +1,2 @@
-vinmonopolet/updateprogram/Main.java
+vinmonopolet/updateprogram/src/main/resources/api_key.txt
+*.class
