+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "maven.view": "hierarchical",
+  "java.configuration.updateBuildConfiguration": "interactive"
+}
