--- vinmonopolet/updateProgram/stores.json
@@ -1,1 +1,0 @@
-[]
