+++ vinmonopolet/updateprogram/src/test/java/com/updateprogram/AppTest.java
@@ -1,0 +1,20 @@
+package com.updateprogram;
+
+import static org.junit.Assert.assertTrue;
+
+import org.junit.Test;
+
+/**
+ * Unit test for simple App.
+ */
+public class AppTest 
+{
+    /**
+     * Rigorous Test :-)
+     */
+    @Test
+    public void shouldAnswerWithTrue()
+    {
+        assertTrue( true );
+    }
+}
